Move Unsubscribable error into its own class
To declutter the line item builder and to have the error belong to the
SolidusSubscriptions namespace
