+++ app/models/solidus_subscriptions/unsubscribable_error.rb
@@ -1,0 +1,17 @@
+# This error should be raised if a user attempts to subscribe to a item which
+# is not subscribable
+module SolidusSubscriptions
+  class UnsubscribableError < StandardError
+    def initialize(subscribable)
+      @subscribable = subscribable
+      super
+    end
+
+    def to_s
+      <<-MSG.squish
+        #{@subscribable.class} with id: #{@subscribable.id} cannot be
+        subscribed to.
+      MSG
+    end
+  end
+end
